# Use maximum zlib compression when generating editor translation headers