Follow-up to https://github.com/godotengine/godot/pull/54386.

With comments stripped (using the above PR), this reduces the combined generated translation size from 28.7 MB to 28.4 MB (-240 KB). Decompression speed is unaffected by the zlib compression level chosen, so the editor startup performance remains identical.

If Zstandard was available in the Python standard library, we could use it, but I don't think it's available there.

This partially addresses https://github.com/godotengine/godot-proposals/issues/3421.

### Before (with https://github.com/godotengine/godot/pull/54386)

```md
15,107,211  editor/doc_translations.gen.h
13,613,454  editor/editor_translations.gen.h
```

### After (with https://github.com/godotengine/godot/pull/54386 + this PR)

```md
15,017,815  editor/doc_translations.gen.h
13,464,278  editor/editor_translations.gen.h
```