@Calinou 2021-10-29
> I think you can use the `zlib.Z_BEST_COMPRESSION` constant instead of the `9`, which may be more readable. Otherwise looks good to me. Any rough idea on the impact on the initial build speed?

Done. I also amended the PR to use the best compression possible in other places where `zlib.compress()` is used.

## Build time comparison

### Before this PR

```
❯ hyperfine -p "rm editor/*translations.gen.h -f" "pyston-scons -j10 platform=linuxbsd use_llvm=yes use_lld=yes werror=no tests=no"
Benchmark #1: pyston-scons -j10 platform=linuxbsd use_llvm=yes use_lld=yes werror=no tests=no
  Time (mean ± σ):     13.854 s ±  0.102 s    [User: 13.691 s, System: 0.986 s]
  Range (min … max):   13.726 s … 13.996 s    10 runs
```

### After this PR

```
❯ hyperfine -p "rm editor/*translations.gen.h -f" "pyston-scons -j10 platform=linuxbsd use_llvm=yes use_lld=yes werror=no tests=no"
Benchmark #1: pyston-scons -j10 platform=linuxbsd use_llvm=yes use_lld=yes werror=no tests=no
  Time (mean ± σ):     13.837 s ±  0.051 s    [User: 14.144 s, System: 0.988 s]
  Range (min … max):   13.747 s … 13.906 s    10 runs
```

### Null build

```
❯ hyperfine "pyston-scons -j10 platform=linuxbsd use_llvm=yes use_lld=yes werror=no tests=no"           Benchmark #1: pyston-scons -j10 platform=linuxbsd use_llvm=yes use_lld=yes werror=no tests=no
  Time (mean ± σ):     12.190 s ±  0.098 s    [User: 11.275 s, System: 0.917 s]
  Range (min … max):   12.095 s … 12.373 s    10 runs
```