@akien-mga 2021-11-01
> Done. I also amended the PR to use the best compression possible in other places where `zlib.compress()` is used.

Seems like you forgot to push that amend.